requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
marko==1.3.0
multidict==6.0.4
openai==0.27.4
Pillow==9.5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
python-pptx==0.6.21
regex==2023.3.23
requests==2.28.2
soupsieve==2.4.1
tqdm==4.65.0
urllib3==1.26.15
XlsxWriter==3.1.0
yarl==1.8.2
